backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Marcus Schiesser <mail@marcusschiesser.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11,<3.12"
fastapi = "^0.104.1"
uvicorn = { extras = ["standard"], version = "^0.23.2" }
llama-index = "^0.9.19"
pypdf = "^3.17.0"
python-dotenv = "^1.0.0"
transformers = "^4.36.2"
torch = "^2.1.2"
ebooklib = "^0.18"
html2text = "^2020.1.16"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
